BuxferConnector/requirements.txt
azure-cosmos
pylint
urllib3